+++ resources/views/components/checkbox.blade.php
@@ -1,0 +1,1 @@
+<input type="checkbox" {{ $attributes->merge(['class' => 'h-4 w-4 rounded border-ink-700 bg-ink-850 text-brand-600 focus:ring-2 focus:ring-brand-500/30 focus:ring-offset-0']) }}>